feat: Create wallet page
created the wallet page

Davi Konuma

+++ backend/userApi/Application/Services/UserService.cs
@@ -1,6 +1,4 @@
-using Microsoft.AspNetCore.Http.HttpResults;
-
-public class UserService : IUserService
+﻿public class UserService : IUserService
 {
     private readonly IUserRepository _userRepository;
 
@@ -13,9 +11,9 @@ public class UserService : IUserService
     {
         var hashedPassword = BCrypt.Net.BCrypt.HashPassword(userDto.Password);
 
-        var user = new User 
-        { 
-            Name = userDto.Name, 
+        var user = new User
+        {
+            Name = userDto.Name,
             Email = userDto.Email,
             Phone = userDto.Phone,
             Address = userDto.Address,
@@ -38,15 +36,17 @@ public class UserService : IUserService
     public UserDTO? GetUserDetails(int id)
     {
         var user = _userRepository.GetById(id);
-        return user != null ? new UserDTO 
-        { 
-            Name = user.Name, 
-            Email = user.Email,
-            Phone = user.Phone,
-            Address = user.Address,
-            Password = user.Password,
-            Photo = user.Photo
-        } : null;
+        return user != null
+            ? new UserDTO
+            {
+                Name = user.Name,
+                Email = user.Email,
+                Phone = user.Phone,
+                Address = user.Address,
+                Password = user.Password,
+                Photo = user.Photo
+            }
+            : null;
     }
 
     public UserDTO[] GetAllUsers()
@@ -54,12 +54,12 @@ public class UserService : IUserService
         var users = _userRepository.ListAll();
         var userDTOs = new List<UserDTO>();
 
-        foreach(var user in users)
+        foreach (var user in users)
         {
-            userDTOs.Add(new UserDTO 
-            { 
-                Id = user.Id, 
-                Name = user.Name, 
+            userDTOs.Add(new UserDTO
+            {
+                Id = user.Id,
+                Name = user.Name,
                 Email = user.Email,
                 Phone = user.Phone,
                 Address = user.Address,
@@ -70,8 +70,8 @@ public class UserService : IUserService
 
         return userDTOs.ToArray();
     }
-    
-     public UpdateUserDTO? UpdateUser(int id, UpdateUserDTO userDto)
+
+    public UpdateUserDTO? UpdateUser(int id, UpdateUserDTO userDto)
     {
         var user = _userRepository.GetById(id);
         if (user == null)
@@ -79,51 +79,64 @@ public class UserService : IUserService
             return null;
         }
 
-        // Atualiza apenas os campos que foram fornecidos
         if (userDto.Name != null)
+        {
             user.Name = userDto.Name;
-            
-        if (userDto.Email != null)
-            user.Email = userDto.Email;
-            
-        if (userDto.Phone != null)
-            user.Phone = userDto.Phone;
-            
-        if (userDto.Address != null)
-            user.Address = userDto.Address;
+        }
 
-        // Atualiza a senha apenas se fornecida
+        if (userDto.Email != null)
+        {
+            user.Email = userDto.Email;
+        }
+
+        if (userDto.Phone != null)
+        {
+            user.Phone = userDto.Phone;
+        }
+
+        if (userDto.Address != null)
+        {
+            user.Address = userDto.Address;
+        }
+
         if (!string.IsNullOrEmpty(userDto.Password))
         {
             user.Password = BCrypt.Net.BCrypt.HashPassword(userDto.Password);
         }
 
         if (userDto.Photo != null)
+        {
             user.Photo = userDto.Photo;
+        }
 
         _userRepository.Update(user);
 
         return new UpdateUserDTO
         {
+            Id = user.Id,
             Name = user.Name,
             Email = user.Email,
             Phone = user.Phone,
             Address = user.Address,
             Photo = user.Photo
-            // Não retornar a senha, mesmo que hasheada
         };
     }
 
     public void DeleteUser(int id)
     {
         var user = _userRepository.GetById(id);
+        if (user == null)
+        {
+            throw new KeyNotFoundException($"Usuário com o id {id} não encontrado.");
+        }
+
         _userRepository.Delete(user);
     }
 
     public UserDTO? ValidateUser(string email, string password)
     {
         var user = _userRepository.GetByEmail(email);
-        if(user == null || !BCrypt.Net.BCrypt.Verify(password, user.Password))
+        if (user == null || !BCrypt.Net.BCrypt.Verify(password, user.Password))
         {
             return null;
         }
